Fix assert: run integration server in test process
With the newer Envoy we structurally run into an assert in
the integration test because it doesn't like simulated time
advancing on a different thread id. To fix that, we run the
integration test server in the parent process and fork of
our own tests into a different process, which is the opposite
of how that worked before this change.

Signed-off-by: Otto van der Schaaf <oschaaf@we-amp.com>